requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
google-genai>=1.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
plotly>=5.20.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyarrow>=15.0.0
tabulate>=0.9.0
